% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addconditions}
\alias{addconditions}
\title{Create condition variables and add them to the dataset}
\usage{
addconditions(data, spring = 15:26, autumn = 36:45, duration = 5 * 52,
  last = NULL, delay = 0, group = NULL)
}
\arguments{
\item{data}{input data in EuroMOMO format}

\item{group}{which group to use. Groups are defined using variables so this must be a name of an actual variable in the data}

\item{spring:}{Week numbers for spring period, vector of integers between 1 and 53}

\item{autumn:}{Week numbers for autumn period, vector of integers between 1 and 53}

\item{duration:}{Duration of the baseline}

\item{last:}{The last period that will be excluded}

\item{delay:}{the number of delay week}
}
\value{
EuroMOMO data with extra variables with conditions used for modelling
}
\description{
Create condition variables and add them to the dataset
}

