% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addMetaData.R
\name{addMetaData}
\alias{addMetaData}
\title{Function to add meta data.}
\usage{
addMetaData(df, groupName, groupOptions)
}
\arguments{
\item{df}{A \code{data.frame} containing the model output.}

\item{groupName}{A string containing the name of the group.}

\item{groupOptions}{A vector of strings containing the group specific options.}
}
\description{
Function to add meta data.
}
\note{
There is a lot of redudant data here!
}

