% Generated by roxygen2 (4.0.2): do not edit by hand
\name{addweeks}
\alias{addweeks}
\title{Create week variable and trend variable}
\usage{
addweeks(data, group = NULL)
}
\arguments{
\item{data}{input data in EuroMOMO format}

\item{group}{which group to use. Groups are defined using variables so this must be a name of an actual variable in the data}
}
\value{
EuroMOMO data with extra variables with YearWeek and trend variable
}
\description{
Create week variable and trend variable
}

