% Generated by roxygen2 (4.0.2): do not edit by hand
\name{baseline}
\alias{baseline}
\title{Calculate baseline}
\usage{
baseline(data, groupOptions, seasonality = 1, trend = 1, ...)
}
\arguments{
\item{data}{input data in EuroMOMO format}

\item{groupOptions}{vector of group specific options}

\item{seasonality}{number of seasonality components (0 or 1) (overrides groupOptions, if given)}

\item{trend}{include a linear trend (0 or 1) (overrides groupOptions, if given)}

\item{...}{Extra parameters for glm}
}
\value{
EuroMOMO data with predicted values, prediction variances and overdispersion
}
\description{
Calculate baseline
}

