% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkOptions}
\alias{checkOptions}
\title{Function to check, if the currently list stored in options("euromomo")
is semantically valid. At the moment, this check consists of:
1. Check for all entries in 'except' that dStart <= dEnd
2. Each group has a 'definition' and a 'label' attribute.
3. That all Boolean Attributes (e.g. 'trend' and 'seasonality') are really Booleans.
At the first error the function stops.}
\usage{
checkOptions()
}
\value{
TRUE, if functions passes finds no errors.
}
\description{
Function to check, if the currently list stored in options("euromomo")
is semantically valid. At the moment, this check consists of:
1. Check for all entries in 'except' that dStart <= dEnd
2. Each group has a 'definition' and a 'label' attribute.
3. That all Boolean Attributes (e.g. 'trend' and 'seasonality') are really Booleans.
At the first error the function stops.
}

