% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delay.nb}
\alias{delay.nb}
\title{R function for delay corretion based on negative binomial distribution}
\usage{
delay.nb(rTDF, holiday)
}
\arguments{
\item{rTDF}{dataframe with aggregated number of reported deaths}

\item{holiday}{file with holiday definitions}
}
\value{
a data frame with added column for delay corrected number of deaths and it's estimated variance
}
\description{
R function for delay corretion based on negative binomial distribution
}

