% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delay.poisson}
\alias{delay.poisson}
\title{Delay correction, Poisson version}
\usage{
delay.poisson(rTDF, method = c("poisson", "negbin"), opentype = c("offset",
  "effect", "none"), openvar = c("open", "propopen"),
  delaytype = c("factor", "numeric", "none"), holiday)
}
\arguments{
\item{rTDF}{source data from aggregation (reporting triangle in cumulated form)}

\item{method}{either "poisson" or "negbin"}

\item{opentype}{which type of effect for the "open"?}

\item{openvar}{which version of the variable "open" to user=}

\item{delaytype}{which kind of effect for the delay variable?}

\item{holiday}{file with holiday definitions}
}
\value{
a data frame with added column for delay corrected number of deaths and it's estimated variance
}
\description{
Delay correction, Poisson version
}

