% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delay.R
\name{delay}
\alias{delay}
\title{Delay correction}
\usage{
delay(data, method = c("poisson", "negbin"), ..., holiday)
}
\arguments{
\item{data}{source data from aggregation}

\item{method}{either "poisson" or "negbin"}

\item{...}{other options passed to actual algorithms}

\item{holiday}{file with holiday definitions}
}
\value{
a data frame with added column for delay corrected number of deaths and it's estimated variance
}
\description{
Delay correction
}

