% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diagnostic.plots}
\alias{diagnostic.plots}
\title{R function plotting histograms and scatter plots of data, residuals and Z-scores}
\usage{
diagnostic.plots(data, plot.options = c("matrix", "singles", "none"))
}
\arguments{
\item{data}{is a dataframe with aggregated number of reported deaths, baseline, Zscores}

\item{plot.options}{selects for output graph type, default is matrix}
}
\description{
R function plotting histograms and scatter plots of data, residuals and Z-scores
}

