% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diagostic.plots}
\alias{diagostic.plots}
\title{R function for delay corretion based on negative binomial distribution}
\usage{
diagostic.plots(data, plot.options = c("matrix", "singles", "none"))
}
\arguments{
\item{data}{is a dataframe with aggregated number of reported deaths, baseline, Zscores}

\item{plot.options}{selects for output graph type, default is matrix}
}
\value{
no value as yet
}
\description{
R function for delay corretion based on negative binomial distribution
}

