% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baseline.R
\name{excess}
\alias{excess}
\title{Excess estimation}
\usage{
excess(data, multiplier = 2, type = momoopts("DelayVariance"))
}
\arguments{
\item{data}{a data frame in EuroMoMo format}

\item{multiplier}{how many approximate standard deviations to use?}

\item{type}{Which type to append ('baseline', 'basedelay', 'delay' or 'both') or 'both')}
}
\value{
a data frame in EuroMoMo format
}
\description{
Calculates confidence intervals for the prediction and excess
}

