% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/delay.R
\name{holiday}
\alias{holiday}
\title{Manage the holiday files}
\usage{
holiday(holiday.filename = "data/holidays.txt")
}
\arguments{
\item{holiday.filename}{name of the file}
}
\value{
a data.frame with columns ISOweek and something
}
\description{
Manage the holiday files
}

