% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{ISOseasonStart}
\alias{ISOseasonStart}
\title{Function that calculates the start of the season}
\usage{
ISOseasonStart(x)
}
\arguments{
\item{x}{vector of ISOweek}
}
\value{
a vector of begining of the season
}
\description{
Function that calculates the start of the season
}

