% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ISOwoy}
\alias{ISOwoy}
\title{Function to extract the week of the year (WOY) as decimal number (00–53)
(i.e. not the year, just the number) from an ISOweek::ISOweek generated date.
Example: For 2013-W13 this would be 13.}
\usage{
ISOwoy(x)
}
\arguments{
\item{x}{A vector of ISOweek::ISOweek generated objects.}
}
\value{
Vector of numeric each being between 00-53.
}
\description{
Function to extract the week of the year (WOY) as decimal number (00–53)
(i.e. not the year, just the number) from an ISOweek::ISOweek generated date.
Example: For 2013-W13 this would be 13.
}

