% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/makeGroups.R
\name{makeGroups}
\alias{makeGroups}
\title{Create group columns in the data frame.}
\usage{
makeGroups(momo)
}
\arguments{
\item{momo}{A data.frame containing the data set.}
}
\value{
The modified data.frame
}
\description{
Function to parse the group definitions and put in a column for each age group.
These are evaluated within the data.frame.
}

