% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{na.0}
\alias{na.0}
\title{Simple example function}
\usage{
na.0(a, b = 0)
}
\arguments{
\item{a}{a parameter}

\item{b}{an alternative}
}
\value{
another value
}
\description{
Simple example function
}

