% Generated by roxygen2 (4.0.2): do not edit by hand
\name{output}
\alias{output}
\title{Generate outputs}
\usage{
output(data)
}
\arguments{
\item{data}{\code{data.frame} generated by baseline function}
}
\description{
Generate outputs
}
\section{graphs graphs with crude mortality and z-scores}{

}

\section{table table with cumulative mortality per season}{

}

\section{data text file containing the original dataframe}{

}

