% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{parseDefaultsFile}
\alias{parseDefaultsFile}
\title{Function to parse a EuroMOMO specification file.}
\usage{
parseDefaultsFile(fileName, debug = FALSE)
}
\arguments{
\item{fileName}{of the parameter configuration file}

\item{debug}{if true print extensive information}
}
\description{
The function sets R-EuroMOMO algorithm parameters based on a cascading series of TXT file consisting
of <Lefthandside> = <Righthandside> operators.
}
\details{
Operation: Each <LHS> = <RHS> is stored in a list. There are two special
LHS's: except and group, which are used to specify exception periods and group
information. First a global defaults.txt file is ran and then possibly local
defaults are run. Parameters which need to be here: see ImportantVarNames in checkOptions.\cr\cr
Syntax for specifying groups: \code{<group>.<groupName>.<attributeName> = <attributeValue}. \cr
Syntax for specifying except periods: \code{except = <Start ISO-week>:<End ISO-week>}. \cr \cr
There are five pre-defined age groups: momodefault<1-5>.
}
\note{
The parser is handwritten based on regular expressions, which is prone to error.
Future versions could be, e.g., XML based.
}

