% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotRTs.R
\name{plotDelayDiagnostics2File}
\alias{plotDelayDiagnostics2File}
\title{A function combining the plotDelay and plotDelay quantile function}
\usage{
plotDelayDiagnostics2File(rTList, w = 1, quantiles = c(0.25, 0.5, 0.75, 0.9,
  0.95, 0.99), main = NULL, week.dir)
}
\arguments{
\item{rTList}{List containing different versions of the reporting triangle data.frame}

\item{w}{Size of the moving window used for smoothing}

\item{quantiles}{Vector containing the quantiles to show in the 2nd plot of the delay distribution.}

\item{main}{Title of the graphics (typically the label of the age group)}

\item{week.dir}{Directory name where all output is stored.}
}
\description{
A function combining the plotDelay and plotDelay quantile function
}

