% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/diagnositics.R
\name{plothist}
\alias{plothist}
\title{function for plotting histograms, called from within the main diagnostic.plots}
\usage{
plothist(x, plotline = FALSE, title = NULL)
}
\arguments{
\item{x}{A vector of values for the histogram}

\item{plotline}{TRUE or FALSE (default) for plotting normal fit line}

\item{title}{to pass as the title in a plot}
}
\description{
function for plotting histograms, called from within the main diagnostic.plots
}

