% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotscatter}
\alias{plotscatter}
\title{function for scatter plots, called from within the main diagnostic.plots}
\usage{
plotscatter(x, y, plotline = TRUE, title = NULL)
}
\arguments{
\item{x}{A x-axis values for the scatter plot}

\item{y}{A y-axis values for the scatter plot}

\item{plotline}{TRUE (default) or FALSE for plotting slr fit line}

\item{title}{to pass as the title in a plot}
}
\description{
function for scatter plots, called from within the main diagnostic.plots
}

