% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotspectrum}
\alias{plotspectrum}
\title{function for plotting periodograms (spectral densities on the log scale), called from within the main diagnostic.plots}
\usage{
plotspectrum(x, plotline = TRUE, title = NULL)
}
\arguments{
\item{x}{A vector of values from a time series}

\item{plotline}{TRUE (default) or FALSE for plotting vertical lines}

\item{title}{to pass as the title in a plot}
}
\description{
function for plotting periodograms (spectral densities on the log scale), called from within the main diagnostic.plots
}

