% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/reader.R
\name{readmomofile}
\alias{readmomofile}
\title{Function to read a file with individual case data from file.}
\usage{
readmomofile(dateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{dateFormat}{Format string specifying how the dates are formatted. Default: \code{"\%Y-\%m-\%d"}.}
}
\value{
Returns a list containing the elements
\tabular{ll}{
momo \tab \code{data.frame} with all the relevant data. Note that only numberOfSeasons are kept. \cr
dWeeks \tab Vector of mondays corresponding to the weeks spanning the data \cr
dLastFull \tab Monday of the last full week in the data.
}
}
\description{
The file needs to be a CSV file with at least columns the columns:
DoD (Day of Death) and DoR (Day of Registration) and age (Age of individual who died).
All Dates need to be in ISO 8601 format, i.e. YYYY-MM-DD.
}

