% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rT2DataFrame}
\alias{rT2DataFrame}
\title{Turn a triangle into a data.frame}
\usage{
rT2DataFrame(rT)
}
\arguments{
\item{rT}{an output from df2ReportingTriangle}
}
\value{
a data frame with column ISOweek and one column for each delay band
}
\description{
Turn a triangle into a data.frame
}

