% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotRTs.R
\name{verticalLineSDE}
\alias{verticalLineSDE}
\title{Draw vertical line indicating start of the delay estimation.}
\usage{
verticalLineSDE(mondays)
}
\arguments{
\item{mondays}{Vector of dates indicating the mondays of the ISOweeks.}
}
\description{
This internal helper function draws a vertical line and adds an axis label
at the top of the graph.
}

