% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/output.R
\name{writeHUBFiles}
\alias{writeHUBFiles}
\title{Write final output to HUB compatible files.}
\usage{
writeHUBFiles(final, dLastFullWeek)
}
\arguments{
\item{final}{Final data.frame}

\item{dLastFullWeek}{Last final week.}
}
\value{
Nothing
}
\description{
Write final output to HUB compatible files.
}

