% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baseline.R
\name{addweeks}
\alias{addweeks}
\title{Create week variable and trend variable}
\usage{
addweeks(data)
}
\arguments{
\item{data}{input data in EuroMOMO format}
}
\value{
EuroMOMO data with extra variables with YearWeek and trend variable
}
\description{
Create week variable and trend variable
}

