% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkOptions}
\alias{checkOptions}
\title{Checking of the euromomo options.}
\usage{
checkOptions()
}
\value{
TRUE, if function finds no errors. Otherwise a "stop" halts function execution.
}
\description{
Function to check, if the currently list stored in options("euromomo")
is semantically valid.
}
\details{
At the moment, this check consists of:
\enumerate{
\item Check for all entries in 'except' that dStart <= dEnd
\item Each group has a 'definition' and a 'label' attribute.
\item That all Boolean Attributes (e.g. 'trend' and 'seasonality') are really Booleans.
At the first error the function stops.
}
}
\note{
Only one error at the time is found.
}

