% Generated by roxygen2 (4.0.2): do not edit by hand
\name{delay}
\alias{delay}
\title{Delay correction}
\usage{
delay(data, method = c("poisson", "negbin"), ..., holiday)
}
\arguments{
\item{method}{either "poisson" or "negbin"}

\item{...}{other options passed to actual algorithms}

\item{holiday}{file with holiday definitions}

\item{rTDF}{source data from aggregation}

\item{a}{data frame with added column for delay corrected number of deaths and it's estimated variance}
}
\description{
Delay correction
}

