% Generated by roxygen2 (4.0.2): do not edit by hand
\name{df2ReportingTriangle}
\alias{df2ReportingTriangle}
\title{R function for reading in and performing the aggregation necessary
to get the reporting triangle.}
\usage{
df2ReportingTriangle(momo, groupindicator, back, dWeeks, dLastFullWeek)
}
\arguments{
\item{momo}{is a data.frame with the complete data
back is the number of weeks for delay-adjustment age group specific
groupindicator is the indicator of the age group for data aggregation}
}
\value{
A list containing the reporting triangle for each age group, the cumulated reporting triangle, the time points and the delays.
}
\description{
R function for reading in and performing the aggregation necessary
to get the reporting triangle.
}

