% Generated by roxygen2 (4.0.2): do not edit by hand
\name{excess}
\alias{excess}
\title{Excess estimation}
\usage{
excess(data, multiplier = 2, type = momoopts("DelayVariance"))
}
\arguments{
\item{data}{a data frame in EuroMoMo format}

\item{multiplier}{how many approximate standard deviations to use?}
}
\value{
a data frame in EuroMoMo format
}
\description{
Calculates confidence intervals for the prediction and excess
}

