% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ISOseasonStart}
\alias{ISOseasonStart}
\title{Function that calculates the start of the season}
\usage{
ISOseasonStart(x)
}
\arguments{
\item{x:}{vector of ISOweek}
}
\value{
a vector of begining of the season
}
\description{
Function that calculates the start of the season
}

