% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{ISOyear}
\alias{ISOyear}
\title{Function to extract the year of a YYYY-WXX representation.}
\usage{
ISOyear(x)
}
\arguments{
\item{x}{A vector of \code{ISOweek::ISOweek} generated objects.}
}
\value{
Vector of numeric containing the year.
}
\description{
Extract year from a vector of \code{ISOweek::ISOweek} generated dates.
Example: For 2013-W13 this would be 2013.
}

