% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeGroups}
\alias{makeGroups}
\title{Function to parse the group definitions and put in a column for each age group.
These are evaluated within the data.frame.}
\usage{
makeGroups(momo)
}
\arguments{
\item{momo}{A data.frame containing the data set.}
}
\value{
The modified data.frame
}
\description{
Function to parse the group definitions and put in a column for each age group.
These are evaluated within the data.frame.
}

