% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/output.R
\name{output}
\alias{output}
\title{Generate outputs from the EuroMomo algorithm}
\usage{
output(data)
}
\arguments{
\item{data}{\code{data.frame} generated by baseline function}
}
\description{
The function writes output graph files, tables into a textfile, etc.
}
\section{graphs}{

graphs with crude mortality and z-scores
}

\section{table}{

table with cumulative mortality per season
}

\section{data}{

text file containing the original dataframe
}

