% Generated by roxygen2 (4.0.2): do not edit by hand
\name{parseDefaultsFile}
\alias{parseDefaultsFile}
\title{Function to parse a specification file for setting the
R-EuroMOMO algorithm parameters. The parser is handwritten based
on regular expressions. This is error prone. Future versions could
be, e.g., XML based, but this works for now. Based on the philosophy
of a SINGLE parameter file, we removed any previous default configurations,
but this would be easy to add again (e.g. one defining the 4 STANDARD age groups.)}
\usage{
parseDefaultsFile(fileName, debug = FALSE)
}
\arguments{
\item{fileName}{of the parameter configuration file}

\item{debug}{if true print extensive information}
}
\description{
Function to parse a specification file for setting the
R-EuroMOMO algorithm parameters. The parser is handwritten based
on regular expressions. This is error prone. Future versions could
be, e.g., XML based, but this works for now. Based on the philosophy
of a SINGLE parameter file, we removed any previous default configurations,
but this would be easy to add again (e.g. one defining the 4 STANDARD age groups.)
}

