% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotRTs.R
\name{plotDelay}
\alias{plotDelay}
\title{Show the delay as a function of time.}
\usage{
plotDelay(df, main = NULL)
}
\arguments{
\item{df}{a data frame representing a reporting triangle (Note: it's cumulative).}

\item{main}{Title of the plot (typically the label of the agegroup)}
}
\value{
nothing (NULL)
}
\description{
Show the delay as a function of time.
}

