% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDelay}
\alias{plotDelay}
\title{Show the delay as a function of time.}
\usage{
plotDelay(df)
}
\arguments{
\item{df}{a data frame representing a reporting triangle.}
}
\value{
nothing (NULL)
}
\description{
Show the delay as a function of time.
}

