% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDelayDiagnostics2File}
\alias{plotDelayDiagnostics2File}
\title{A function combining the plotDelay and plotDelay quantile function}
\usage{
plotDelayDiagnostics2File(rT, w = 1, main = NULL, week.dir)
}
\arguments{
\item{rT}{Reporting triangle data.frame}

\item{w}{Size of the moving window used for smoothing}

\item{main}{Title of the graphics (typically the label of the age group)}

\item{week.dir}{Directory name where all output is stored.}
}
\description{
A function combining the plotDelay and plotDelay quantile function
}

