% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotDelayQuantiles}
\alias{plotDelayQuantiles}
\title{Illustrate empirical delay distribution.}
\usage{
plotDelayQuantiles(rT, w = 1, ISOweeks, quantiles = c(0.1, 0.5, 0.9),
  col = 1, lty = 1, lwd = 2, main = NULL)
}
\arguments{
\item{rT}{- reporting triangle as it would be at the end.}

\item{w}{- half-width of moving window}

\item{ISOweeks}{Vector of ISO week dates where to show the result}

\item{quantiles}{- which quantiles to show}

\item{col}{Color of the quantiles, see \code{\link{plot}}.}

\item{lty}{Line types, see \code{\link{plot}}.}

\item{lwd}{Line width, see \code{\link{plot}}.}

\item{main}{- Title of the plot (tpyically the label of the age group)}
}
\value{
Nothing
}
\description{
Show empirical and model based median of delay distribution as a
function of occurence time t.
}

