% Generated by roxygen2 (4.0.2): do not edit by hand
\name{writeHUBFiles}
\alias{writeHUBFiles}
\title{Write final output to HUB compatible files.}
\usage{
writeHUBFiles(final, dLastFullWeek)
}
\arguments{
\item{final}{Final data.frame}

\item{dLastFullWeek}{Last final week.}
}
\value{
Nothing
}
\description{
Write final output to HUB compatible files.
}

