% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zscore}
\alias{zscore}
\title{Append zscores to the EuroMoMo data}
\usage{
zscore(data, type = momoopts("DelayVariance"))
}
\arguments{
\item{data}{a data frame in EuroMoMo format}
}
\value{
a data frame in EuroMoMo format
}
\description{
Append zscores to the EuroMoMo data
}

