% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/baseline.R
\name{zscore}
\alias{zscore}
\title{Append zscores to the EuroMoMo data}
\usage{
zscore(data, type = momoopts("DelayVariance"))
}
\arguments{
\item{data}{a data frame in EuroMoMo format}

\item{type}{Which type to append ('baseline' or 'both')}
}
\value{
a data frame in EuroMoMo format
}
\description{
Append zscores to the EuroMoMo data
}

